requirements-dev.txt
pytest>=6.2.0
tokenize-rt>=4.1.0
mypy>=0.800

# extras dependencies
pyyaml
